Extract Kometa root/install management into modules/kometa_install.py
Phase 4 of breaking up the quickstart.py monolith. This phase turned
out larger than the route count suggested: the 6 routes
(save-kometa-install-mode, validate-kometa-root, probe-kometa-root,
check-kometa-update, update-kometa, update-kometa-progress) sit on top
of a ~500-line "Kometa install selection" service used much more
broadly — by the dependency/validation-summary engine
(_derive_step_status), the /step renderer, and logscan/trends, none of
which are being moved yet. That service layer moves to
modules/kometa_install.py; the routes move to
blueprints/kometa_updates.py.

Re-exported under their original names since they're still called
from quickstart.py: KOMETA_INSTALL_MODE_MANAGED/EXTERNAL,
_canonicalize_kometa_section, _validate_saved_kometa_selection,
_build_kometa_install_context (also monkeypatched directly by
test_workspace_dependency_logic.py), _get_kometa_settings_section,
_resolve_kometa_selection. KOMETA_INSTALL_MODE_EXISTING has no
remaining caller in quickstart.py, so it's only imported where used.

Three more small utilities had no quickstart-specific state and are
also used by not-yet-extracted code (imagemaid), so they moved to
shared modules with the same re-export treatment:
_resolve_user_dir -> helpers.resolve_user_dir,
_retrieve_settings_for_config -> persistence.retrieve_settings_for_config,
apply_validation_metadata -> persistence.apply_validation_metadata.

update-kometa is the one route here that depends on
_get_active_work_blocker, which deliberately stayed in quickstart.py
during the background-jobs extraction (a test monkeypatches one of
its dependencies, and moving the composition function would have
silently broken that patch). Rather than force process-control
concerns into this phase, the route does a deferred `import quickstart`
inside the function body to reach it — a one-line bridge until the
process-control phase extracts _get_active_work_blocker properly.

Found and removed one dead context processor while reading this
cluster: inject_kometa_root was defined but never registered with
@app.context_processor or otherwise called anywhere.

Net: quickstart.py shrinks by 1297 lines.
