+++ quickstart.py
@@ -65,7 +65,17 @@ from modules.background_jobs import (
 )
 from blueprints.validation_routes import bp as validation_routes_bp, refresh_plex_libraries
 from blueprints.asset_routes import bp as asset_routes_bp
+from blueprints.kometa_updates import bp as kometa_updates_bp
 from modules.assets import build_preview_image_data as _build_preview_image_data, list_overlay_fonts
+from modules.kometa_install import (
+    KOMETA_INSTALL_MODE_MANAGED,
+    KOMETA_INSTALL_MODE_EXTERNAL,
+    canonicalize_kometa_section as _canonicalize_kometa_section,
+    validate_saved_kometa_selection as _validate_saved_kometa_selection,
+    build_kometa_install_context as _build_kometa_install_context,
+    get_kometa_settings_section as _get_kometa_settings_section,
+    resolve_kometa_selection as _resolve_kometa_selection,
+)
 
 Request.max_form_parts = 100000  # Allow more form fields if needed
 
@@ -73,6 +83,9 @@ _resolve_request_config_name = persistence.resolve_request_config_name
 utc_now_iso = helpers.utc_now_iso
 _safe_join = helpers.safe_join
 _safe_rel_path = helpers.safe_rel_path
+_resolve_user_dir = helpers.resolve_user_dir
+_retrieve_settings_for_config = persistence.retrieve_settings_for_config
+apply_validation_metadata = persistence.apply_validation_metadata
 
 ACTIVE_WORK_POLICIES = {
     "kometa_run": [
@@ -397,18 +410,6 @@ def _is_valid_auto_sort_hubs_value(value):
     if normalized is None:
         return True
     return normalized in SETTINGS_AUTO_SORT_HUBS_VALUES
-
-
-def apply_validation_metadata(stored_data, status, reason=None, details=None, updated_at=None):
-    if not isinstance(stored_data, dict):
-        stored_data = {}
-    stored_data["validation_status"] = status
-    if reason is not None:
-        stored_data["validation_reason"] = reason
-    if details is not None:
-        stored_data["validation_details"] = details
-    stored_data["validation_updated_at"] = updated_at or utc_now_iso()
-    return stored_data
 
 
 def build_validation_summary(errors):
@@ -3916,28 +3917,6 @@ def _normalize_config_filename(config_name: str | None) -> str:
     return name or "default"
 
 
-def _resolve_user_dir(raw_path: str | None) -> Path | None:
-    if not isinstance(raw_path, str):
-        return None
-    raw_path = raw_path.strip()
-    if not raw_path:
-        return None
-    if "\x00" in raw_path:
-        return None
-    try:
-        path = Path(raw_path)
-    except Exception:
-        return None
-    if not path.is_absolute():
-        return None
-    if any(part == ".." for part in path.parts):
-        return None
-    try:
-        return path.resolve()
-    except Exception:
-        return None
-
-
 def _rename_config_files(old_name: str, new_name: str, dry_run: bool = False) -> dict:
     result = {"success": False, "renamed": [], "skipped": [], "errors": [], "rollback_errors": []}
     old_norm = _normalize_config_filename(old_name)
@@ -4058,19 +4037,6 @@ DOTENV = os.path.relpath(os.path.join(helpers.CONFIG_DIR, ".env"))
 load_dotenv(DOTENV, override=True)
 
 
-def _config_name_from_yaml_filename(filename: str | None) -> str | None:
-    text = str(filename or "").strip()
-    if not text:
-        return None
-    base = Path(text).name
-    lowered = base.lower()
-    if lowered.endswith("_config.yml"):
-        return base[:-11]
-    if lowered.endswith("_config.yaml"):
-        return base[:-12]
-    return None
-
-
 # Initialize logging
 helpers.initialize_logging()
 
@@ -4084,6 +4050,7 @@ app = Flask(__name__)
 
 app.register_blueprint(validation_routes_bp)
 app.register_blueprint(asset_routes_bp)
+app.register_blueprint(kometa_updates_bp)
 
 # Run version check at startup
 app.config["VERSION_CHECK"] = helpers.check_for_update()
@@ -4092,10 +4059,6 @@ app.config["VERSION_CHECK"] = helpers.check_for_update()
 kometa_path = os.path.abspath(os.path.join(helpers.CONFIG_DIR, "kometa"))
 
 app.config["KOMETA_ROOT"] = os.environ.get("QS_KOMETA_PATH", kometa_path)
-
-KOMETA_INSTALL_MODE_MANAGED = "managed"
-KOMETA_INSTALL_MODE_EXISTING = "existing"
-KOMETA_INSTALL_MODE_EXTERNAL = "external"
 
 
 def start_update_thread():
@@ -4117,483 +4080,6 @@ def inject_version_info():
         "version_info": app.config.get("VERSION_CHECK") or {},
         "overlay_fonts": list_overlay_fonts(),
     }
-
-
-def inject_kometa_root():
-    return {"kometa_root": app.config["KOMETA_ROOT"]}
-
-
-def _default_managed_kometa_root() -> Path:
-    return Path(helpers.CONFIG_DIR).resolve() / "kometa"
-
-
-def _normalize_kometa_install_mode(value) -> str:
-    raw = str(value or "").strip().lower()
-    if raw == KOMETA_INSTALL_MODE_EXISTING:
-        return KOMETA_INSTALL_MODE_EXISTING
-    if raw == KOMETA_INSTALL_MODE_EXTERNAL:
-        return KOMETA_INSTALL_MODE_EXTERNAL
-    return KOMETA_INSTALL_MODE_MANAGED
-
-
-def _canonicalize_kometa_section(section_data):
-    section = dict(section_data) if isinstance(section_data, dict) else {}
-    canonical = {
-        "install_mode": KOMETA_INSTALL_MODE_MANAGED,
-        "existing_root": "",
-        "external_config_root": "",
-        "external_log_root": "",
-    }
-    canonical.update(section)
-    canonical["install_mode"] = _normalize_kometa_install_mode(canonical.get("install_mode"))
-    canonical["existing_root"] = str(canonical.get("existing_root") or "").strip()
-    canonical["external_config_root"] = str(canonical.get("external_config_root") or "").strip()
-    canonical["external_log_root"] = str(canonical.get("external_log_root") or "").strip()
-    if canonical["install_mode"] != KOMETA_INSTALL_MODE_EXISTING:
-        canonical["existing_root"] = ""
-    if canonical["install_mode"] != KOMETA_INSTALL_MODE_EXTERNAL:
-        canonical["external_config_root"] = ""
-        canonical["external_log_root"] = ""
-    return canonical
-
-
-def _resolve_kometa_selection(section_data=None):
-    section = _canonicalize_kometa_section(section_data)
-    managed_root = _default_managed_kometa_root().resolve()
-    managed_config_dir = managed_root / "config"
-    managed_log_dir = managed_config_dir / "logs"
-    mode = section["install_mode"]
-    existing_root_raw = section.get("existing_root", "")
-    external_config_root_raw = section.get("external_config_root", "")
-    external_log_root_raw = section.get("external_log_root", "")
-    selected_root = managed_root
-    primary_path = managed_root
-    config_dir = managed_config_dir
-    log_dir = managed_log_dir
-    selection_valid = True
-    status_message = "Quickstart will create and manage its own Kometa install under this workspace."
-    mode_label = "Quickstart-managed install"
-    can_launch = True
-    can_update = True
-    can_probe_runtime = True
-    can_read_logs = True
-    can_sync_config = True
-    can_sync_managed_artifacts = True
-
-    if mode == KOMETA_INSTALL_MODE_EXISTING:
-        selection_valid = False
-        status_message = (
-            "Quickstart will use an existing Kometa install visible from this environment. Quickstart will not update that install; update it manually outside Quickstart."
-        )
-        mode_label = "Existing direct install"
-        can_update = False
-        if existing_root_raw:
-            resolved_existing = _resolve_user_dir(existing_root_raw)
-            if resolved_existing:
-                selected_root = resolved_existing
-                primary_path = resolved_existing
-                config_dir = resolved_existing / "config"
-                log_dir = config_dir / "logs"
-                selection_valid = True
-                if not resolved_existing.exists():
-                    status_message = "The selected existing Kometa path is saved, but it is not currently visible from this Quickstart environment."
-                else:
-                    status_message = (
-                        "Quickstart will use the selected existing Kometa install. Quickstart can validate and launch it, but updates must be done manually outside Quickstart."
-                    )
-            else:
-                selected_root = None
-                primary_path = None
-                config_dir = None
-                log_dir = None
-                status_message = "The saved existing Kometa path is invalid."
-        else:
-            selected_root = None
-            primary_path = None
-            config_dir = None
-            log_dir = None
-            status_message = "Choose an existing Kometa path for this config before using the Kometa run page."
-    elif mode == KOMETA_INSTALL_MODE_EXTERNAL:
-        selected_root = None
-        primary_path = None
-        config_dir = None
-        log_dir = None
-        selection_valid = False
-        mode_label = "External/containerized config+logs"
-        can_launch = False
-        can_update = False
-        can_probe_runtime = False
-        can_sync_config = True
-        can_sync_managed_artifacts = True
-        status_message = "Quickstart will manage config and optional logs for an external Kometa runtime. It will not launch or update Kometa directly."
-        if external_config_root_raw:
-            resolved_config = _resolve_user_dir(external_config_root_raw)
-            if resolved_config:
-                primary_path = resolved_config
-                config_dir = resolved_config
-                selection_valid = True
-                if external_log_root_raw:
-                    resolved_log = _resolve_user_dir(external_log_root_raw)
-                    if resolved_log:
-                        log_dir = resolved_log
-                    else:
-                        selection_valid = False
-                else:
-                    log_dir = resolved_config / "logs"
-                can_read_logs = bool(log_dir)
-                if not resolved_config.exists():
-                    status_message = "The selected external Kometa config path is saved, but it is not currently visible from this Quickstart environment."
-                else:
-                    status_message = "Quickstart will sync generated config and managed artifacts to the selected external Kometa config path."
-                if log_dir and not log_dir.exists():
-                    can_read_logs = False
-                    status_message += " Log viewing and logscan will stay limited until the log path is accessible."
-            else:
-                status_message = "The saved external Kometa config path is invalid."
-        else:
-            status_message = "Choose the external Kometa config path for this config before using the Kometa run page."
-
-    return {
-        "install_mode": mode,
-        "existing_root": existing_root_raw,
-        "external_config_root": external_config_root_raw,
-        "external_log_root": external_log_root_raw,
-        "managed_root": managed_root,
-        "managed_config_dir": managed_config_dir,
-        "managed_log_dir": managed_log_dir,
-        "selected_root": selected_root,
-        "primary_path": primary_path,
-        "config_dir": config_dir,
-        "log_dir": log_dir,
-        "selection_valid": selection_valid,
-        "status_message": status_message,
-        "is_managed": mode == KOMETA_INSTALL_MODE_MANAGED,
-        "is_external": mode == KOMETA_INSTALL_MODE_EXTERNAL,
-        "mode_label": mode_label,
-        "can_launch": can_launch,
-        "can_update": can_update,
-        "can_probe_runtime": can_probe_runtime,
-        "can_read_logs": can_read_logs,
-        "can_sync_config": can_sync_config,
-        "can_sync_managed_artifacts": can_sync_managed_artifacts,
-    }
-
-
-def _get_kometa_settings_section(config_name=None):
-    resolved_config = config_name or persistence.ensure_session_config_name()
-    settings = _retrieve_settings_for_config(resolved_config, "900-kometa") or {}
-    section = settings.get("kometa", {}) if isinstance(settings, dict) else {}
-    return settings, _canonicalize_kometa_section(section)
-
-
-def _apply_kometa_selection(selection):
-    selection = selection if isinstance(selection, dict) else {}
-    mode = _normalize_kometa_install_mode(selection.get("install_mode"))
-    selected_root = selection.get("selected_root")
-    config_dir = selection.get("config_dir")
-    log_dir = selection.get("log_dir")
-    resolved_root = Path(selected_root).resolve() if selected_root else None
-    resolved_config_dir = Path(config_dir).resolve() if config_dir else None
-    resolved_log_dir = Path(log_dir).resolve() if log_dir else None
-    fallback_root = None
-    if resolved_root is None and resolved_config_dir is not None:
-        fallback_root = resolved_config_dir.parent if resolved_config_dir.name.lower() == "config" else resolved_config_dir
-
-    if has_request_context():
-        session["kometa_install_mode"] = mode
-        session["kometa_root"] = resolved_root.as_posix() if resolved_root else (fallback_root.as_posix() if fallback_root else "")
-        session["kometa_config_dir"] = resolved_config_dir.as_posix() if resolved_config_dir else ""
-        session["kometa_log_dir"] = resolved_log_dir.as_posix() if resolved_log_dir else ""
-
-    if resolved_root:
-        app.config["KOMETA_ROOT"] = str(resolved_root)
-    elif fallback_root:
-        app.config["KOMETA_ROOT"] = str(fallback_root)
-    app.config["KOMETA_INSTALL_MODE"] = mode
-    app.config["KOMETA_CONFIG_DIR"] = str(resolved_config_dir) if resolved_config_dir else ""
-    app.config["KOMETA_LOG_DIR"] = str(resolved_log_dir) if resolved_log_dir else ""
-
-
-def _build_kometa_install_context(config_name=None):
-    _settings, section = _get_kometa_settings_section(config_name)
-    selection = _resolve_kometa_selection(section)
-    _apply_kometa_selection(selection)
-    selected_root = selection.get("selected_root")
-    primary_path = selection.get("primary_path")
-    config_dir = selection.get("config_dir")
-    log_dir = selection.get("log_dir")
-    return {
-        "kometa_install_mode": selection["install_mode"],
-        "kometa_existing_root": selection["existing_root"],
-        "kometa_external_config_root": selection["external_config_root"],
-        "kometa_external_log_root": selection["external_log_root"],
-        "kometa_managed_root": selection["managed_root"].as_posix(),
-        "kometa_managed_root_display": str(selection["managed_root"]),
-        "kometa_managed_config_dir": selection["managed_config_dir"].as_posix(),
-        "kometa_managed_config_dir_display": str(selection["managed_config_dir"]),
-        "kometa_managed_log_dir": selection["managed_log_dir"].as_posix(),
-        "kometa_managed_log_dir_display": str(selection["managed_log_dir"]),
-        "kometa_selected_root": selected_root.as_posix() if selected_root else "",
-        "kometa_selected_root_display": str(selected_root) if selected_root else "",
-        "kometa_primary_path": primary_path.as_posix() if primary_path else "",
-        "kometa_primary_path_display": str(primary_path) if primary_path else "",
-        "kometa_active_config_dir": config_dir.as_posix() if config_dir else "",
-        "kometa_active_config_dir_display": str(config_dir) if config_dir else "",
-        "kometa_active_log_dir": log_dir.as_posix() if log_dir else "",
-        "kometa_active_log_dir_display": str(log_dir) if log_dir else "",
-        "kometa_selection_valid": bool(selection["selection_valid"]),
-        "kometa_install_status_message": selection["status_message"],
-        "kometa_is_managed_install": bool(selection["is_managed"]),
-        "kometa_is_external_install": bool(selection["is_external"]),
-        "kometa_mode_label": selection["mode_label"],
-        "kometa_can_launch": bool(selection["can_launch"]),
-        "kometa_can_update": bool(selection["can_update"]),
-        "kometa_can_probe_runtime": bool(selection["can_probe_runtime"]),
-        "kometa_can_read_logs": bool(selection["can_read_logs"]),
-        "kometa_can_sync_config": bool(selection["can_sync_config"]),
-        "kometa_can_sync_managed_artifacts": bool(selection["can_sync_managed_artifacts"]),
-    }
-
-
-def _resolve_kometa_request_target(payload, logs=None, require_existing_root=False):
-    request_payload = payload if isinstance(payload, dict) else {}
-    config_name = _resolve_request_config_name(request_payload)
-    _settings, stored_section = _get_kometa_settings_section(config_name)
-    requested_mode = request_payload.get("install_mode")
-    install_mode = _normalize_kometa_install_mode(requested_mode if requested_mode not in [None, ""] else stored_section.get("install_mode"))
-    raw_path = str(request_payload.get("path") or "").strip()
-    existing_root = str(request_payload.get("existing_root") or stored_section.get("existing_root") or "").strip()
-    external_config_root = str(request_payload.get("external_config_root") or stored_section.get("external_config_root") or "").strip()
-    external_log_root = str(request_payload.get("external_log_root") or stored_section.get("external_log_root") or "").strip()
-
-    if install_mode == KOMETA_INSTALL_MODE_MANAGED:
-        selection = _resolve_kometa_selection({"install_mode": KOMETA_INSTALL_MODE_MANAGED})
-        if raw_path:
-            resolved_managed = _resolve_user_dir(raw_path)
-            if not resolved_managed:
-                if logs is not None:
-                    logs.append("❌ Invalid path provided.")
-                return {"error": "Invalid path provided."}
-            selection["selected_root"] = resolved_managed
-            selection["primary_path"] = resolved_managed
-            selection["config_dir"] = resolved_managed / "config"
-            selection["log_dir"] = selection["config_dir"] / "logs"
-            selection["selection_valid"] = True
-        _apply_kometa_selection(selection)
-        return {
-            "config_name": config_name,
-            "install_mode": install_mode,
-            "path_obj": selection["selected_root"],
-            "runtime_root": selection["selected_root"],
-            "config_dir": selection["config_dir"],
-            "log_dir": selection["log_dir"],
-            "existing_root": "",
-            "external_config_root": "",
-            "external_log_root": "",
-            "selection_valid": True,
-        }
-
-    if install_mode == KOMETA_INSTALL_MODE_EXTERNAL:
-        candidate = raw_path or external_config_root
-        if not candidate:
-            if logs is not None:
-                logs.append("❌ External Kometa mode requires a config path.")
-            return {"error": "External Kometa mode requires a config path."}
-
-        resolved_config = _resolve_user_dir(candidate)
-        if not resolved_config:
-            if logs is not None:
-                logs.append("❌ Invalid external config path provided.")
-            return {"error": "Invalid external config path provided."}
-
-        resolved_log = None
-        if external_log_root:
-            resolved_log = _resolve_user_dir(external_log_root)
-            if not resolved_log:
-                if logs is not None:
-                    logs.append("❌ Invalid external log path provided.")
-                return {"error": "Invalid external log path provided."}
-
-        if require_existing_root and not resolved_config.exists():
-            if logs is not None:
-                logs.append("❌ The selected external Kometa config path does not exist in this Quickstart environment.")
-            return {"error": "The selected external Kometa config path does not exist in this Quickstart environment."}
-
-        selection = _resolve_kometa_selection(
-            {
-                "install_mode": KOMETA_INSTALL_MODE_EXTERNAL,
-                "external_config_root": candidate,
-                "external_log_root": external_log_root,
-            }
-        )
-        _apply_kometa_selection(selection)
-        return {
-            "config_name": config_name,
-            "install_mode": install_mode,
-            "path_obj": selection["config_dir"],
-            "runtime_root": None,
-            "config_dir": selection["config_dir"],
-            "log_dir": selection["log_dir"],
-            "existing_root": "",
-            "external_config_root": candidate,
-            "external_log_root": external_log_root,
-            "selection_valid": selection["selection_valid"],
-        }
-
-    candidate = raw_path or existing_root
-    if not candidate:
-        if logs is not None:
-            logs.append("❌ Existing Kometa mode requires a path.")
-        return {"error": "Existing Kometa mode requires a path."}
-
-    resolved_existing = _resolve_user_dir(candidate)
-    if not resolved_existing:
-        if logs is not None:
-            logs.append("❌ Invalid path provided.")
-        return {"error": "Invalid path provided."}
-
-    if require_existing_root and not resolved_existing.exists():
-        if logs is not None:
-            logs.append("❌ The selected existing Kometa path does not exist in this Quickstart environment.")
-        return {"error": "The selected existing Kometa path does not exist in this Quickstart environment."}
-
-    selection = _resolve_kometa_selection(
-        {
-            "install_mode": KOMETA_INSTALL_MODE_EXISTING,
-            "existing_root": candidate,
-        }
-    )
-    _apply_kometa_selection(selection)
-    return {
-        "config_name": config_name,
-        "install_mode": install_mode,
-        "path_obj": selection["selected_root"],
-        "runtime_root": selection["selected_root"],
-        "config_dir": selection["config_dir"],
-        "log_dir": selection["log_dir"],
-        "existing_root": candidate,
-        "external_config_root": "",
-        "external_log_root": "",
-        "selection_valid": resolved_existing.exists(),
-    }
-
-
-def _validate_existing_kometa_root(path_obj):
-    p = Path(path_obj).resolve()
-    missing = []
-    if not p.exists():
-        missing.append("path")
-    if not (p / "config").exists():
-        missing.append("config")
-    if not (p / "kometa.py").exists():
-        missing.append("kometa.py")
-    if not (p / "requirements.txt").exists():
-        missing.append("requirements.txt")
-    return missing
-
-
-def _validate_saved_kometa_selection(section_data):
-    section = _canonicalize_kometa_section(section_data)
-    install_mode = section.get("install_mode")
-
-    if install_mode == KOMETA_INSTALL_MODE_MANAGED:
-        return True, None, None
-
-    if install_mode == KOMETA_INSTALL_MODE_EXISTING:
-        existing_root = str(section.get("existing_root") or "").strip()
-        if not existing_root:
-            return False, "missing_location", "Choose the Kometa root folder that contains kometa.py, requirements.txt, and config/."
-        resolved_existing = _resolve_user_dir(existing_root)
-        if not resolved_existing:
-            return False, "invalid_paths", [f"Saved existing Kometa root is invalid. Path: {existing_root}"]
-        if not resolved_existing.exists():
-            return False, "invalid_paths", [f"Existing Kometa root does not exist. Path: {resolved_existing}"]
-        missing = _validate_existing_kometa_root(resolved_existing)
-        if missing:
-            if "path" in missing:
-                return False, "invalid_paths", [f"Existing Kometa root does not exist. Path: {resolved_existing}"]
-            missing_labels = []
-            if "kometa.py" in missing:
-                missing_labels.append("kometa.py")
-            if "requirements.txt" in missing:
-                missing_labels.append("requirements.txt")
-            if "config" in missing:
-                missing_labels.append("config/")
-            label_text = ", ".join(missing_labels) if missing_labels else ", ".join(missing)
-            return False, "invalid_paths", [f"Existing Kometa root is missing required items: {label_text}. Path: {resolved_existing}"]
-        return True, None, None
-
-    external_config_root = str(section.get("external_config_root") or "").strip()
-    external_log_root = str(section.get("external_log_root") or "").strip()
-    if not external_config_root:
-        return False, "missing_location", "Choose the external Kometa config path Quickstart should use for this config."
-    resolved_external_config = _resolve_user_dir(external_config_root)
-    if not resolved_external_config:
-        return False, "invalid_paths", [f"Saved external Kometa config path is invalid. Path: {external_config_root}"]
-    if not resolved_external_config.exists() or not resolved_external_config.is_dir():
-        return False, "invalid_paths", [f"External Kometa config path does not exist. Path: {resolved_external_config}"]
-    if external_log_root:
-        resolved_external_log = _resolve_user_dir(external_log_root)
-        if not resolved_external_log:
-            return False, "invalid_paths", [f"Saved external Kometa log path is invalid. Path: {external_log_root}"]
-        if not resolved_external_log.exists() or not resolved_external_log.is_dir():
-            return False, "invalid_paths", [f"External Kometa log path does not exist. Path: {resolved_external_log}"]
-    return True, None, None
-
-
-def _sync_generated_yaml_and_assets_to_kometa_config(config_dir, config_filename, logs=None):
-    target_config_dir = Path(config_dir).resolve()
-
-    def log(msg):
-        if logs is not None:
-            logs.append(msg)
-
-    config_name = _safe_rel_path(config_filename or "kometa")
-    if not config_name:
-        raise ValueError("Invalid config filename.")
-
-    src_yaml = _safe_join(Path("config"), config_name)
-    if not src_yaml or not src_yaml.exists():
-        raise FileNotFoundError(f"Generated YAML not found: {src_yaml}")
-
-    target_config_dir.mkdir(parents=True, exist_ok=True)
-    dest_yaml = _safe_join(target_config_dir, config_name)
-    if not dest_yaml:
-        raise ValueError("Invalid config destination.")
-
-    shutil.copy2(src_yaml, dest_yaml)
-    log(f"✅ YAML copied to Kometa config folder at: {dest_yaml}")
-
-    yaml_parser = YAML(typ="safe")
-    with src_yaml.open("r", encoding="utf-8") as f:
-        parsed_config = yaml_parser.load(f) or {}
-    active_config_name = session.get("config_name") if has_request_context() else None
-    config_scope = _config_name_from_yaml_filename(config_name) or active_config_name
-    font_refs = helpers.collect_font_references(parsed_config)
-    if font_refs:
-        font_result = helpers.copy_fonts_to_kometa(font_refs, kometa_config_dir=target_config_dir, config_name=config_scope)
-        copied = font_result.get("copied", [])
-        missing = font_result.get("missing", [])
-        errors = font_result.get("errors", [])
-        if copied:
-            log(f"✅ Synced {len(copied)} font(s) referenced in the config to Kometa fonts.")
-        if missing:
-            log(f"⚠️ Fonts referenced in the config not found: {', '.join(missing)}")
-        for err in errors:
-            log(f"⚠️ {err}")
-    if config_scope:
-        artifact_result = helpers.sync_managed_library_artifacts_to_kometa(config_scope, kometa_config_dir=target_config_dir)
-        synced = artifact_result.get("synced", [])
-        removed = artifact_result.get("removed", [])
-        errors = artifact_result.get("errors", [])
-        if synced:
-            log(f"✅ Synced {len(synced)} managed library artifact tree(s) to Kometa config/{config_scope}.")
-        if removed:
-            log(f"ℹ️ Removed {len(removed)} stale managed library artifact tree(s) from Kometa config/{config_scope}.")
-        for err in errors:
-            log(f"⚠️ {err}")
-
-    return {"config_filename": config_name, "destination": dest_yaml}
 
 
 # Use booler() for FLASK_DEBUG conversion
@@ -14902,823 +14388,6 @@ def header_style_previews():
     return jsonify(success=True, previews=previews)
 
 
-@app.route("/save-kometa-install-mode", methods=["POST"])
-def save_kometa_install_mode():
-    payload = request.get_json(silent=True) or {}
-    config_name = _resolve_request_config_name(payload)
-    install_mode = _normalize_kometa_install_mode(payload.get("install_mode"))
-    existing_root = str(payload.get("existing_root") or "").strip()
-    external_config_root = str(payload.get("external_config_root") or "").strip()
-    external_log_root = str(payload.get("external_log_root") or "").strip()
-    managed_root = _default_managed_kometa_root().resolve()
-
-    if install_mode == KOMETA_INSTALL_MODE_EXISTING:
-        if not existing_root:
-            return jsonify(success=False, error="Choose the existing Kometa path Quickstart should use for this config."), 400
-        resolved_existing = _resolve_user_dir(existing_root)
-        if not resolved_existing:
-            return jsonify(success=False, error="The existing Kometa path is invalid."), 400
-        if not resolved_existing.exists():
-            return jsonify(success=False, error="The selected existing Kometa path does not exist in this Quickstart environment."), 400
-        missing = _validate_existing_kometa_root(resolved_existing)
-        if missing:
-            if "kometa.py" in missing or "requirements.txt" in missing or "config" in missing:
-                return (
-                    jsonify(
-                        success=False,
-                        error="Choose the Kometa root folder that contains kometa.py, requirements.txt, and config/.",
-                    ),
-                    400,
-                )
-        section_payload_update = {
-            "install_mode": install_mode,
-            "existing_root": existing_root,
-            "external_config_root": "",
-            "external_log_root": "",
-        }
-    elif install_mode == KOMETA_INSTALL_MODE_EXTERNAL:
-        if not external_config_root:
-            return jsonify(success=False, error="Choose the external Kometa config path Quickstart should use for this config."), 400
-        resolved_external_config = _resolve_user_dir(external_config_root)
-        if not resolved_external_config:
-            return jsonify(success=False, error="The external Kometa config path is invalid."), 400
-        if not resolved_external_config.exists() or not resolved_external_config.is_dir():
-            return jsonify(success=False, error="The selected external Kometa config path does not exist in this Quickstart environment."), 400
-        if external_log_root:
-            resolved_external_log = _resolve_user_dir(external_log_root)
-            if not resolved_external_log:
-                return jsonify(success=False, error="The external Kometa log path is invalid."), 400
-        section_payload_update = {
-            "install_mode": install_mode,
-            "existing_root": "",
-            "external_config_root": external_config_root,
-            "external_log_root": external_log_root,
-        }
-    else:
-        existing_root = ""
-        external_config_root = ""
-        external_log_root = ""
-        section_payload_update = {
-            "install_mode": install_mode,
-            "existing_root": "",
-            "external_config_root": "",
-            "external_log_root": "",
-        }
-
-    stored_validated, stored_user_entered, stored_payload = database.retrieve_section_data(config_name, "kometa")
-    section_payload = stored_payload if isinstance(stored_payload, dict) else {}
-    section_payload["kometa"] = _canonicalize_kometa_section(
-        {
-            **(section_payload.get("kometa") if isinstance(section_payload.get("kometa"), dict) else {}),
-            **section_payload_update,
-        }
-    )
-    section_payload = apply_validation_metadata(section_payload, "validated")
-    database.save_section_data(
-        name=config_name,
-        section="kometa",
-        validated=True,
-        user_entered=True,
-        data=section_payload,
-    )
-    selection = _resolve_kometa_selection(section_payload["kometa"])
-    _apply_kometa_selection(selection)
-
-    if install_mode == KOMETA_INSTALL_MODE_EXISTING:
-        message = "Quickstart will now use the selected existing Kometa install for this config."
-    elif install_mode == KOMETA_INSTALL_MODE_EXTERNAL:
-        message = "Quickstart will now sync config and optional logs for the selected external Kometa setup."
-    else:
-        message = "Quickstart will now use its managed Kometa install for this config."
-    return jsonify(
-        success=True,
-        message=message,
-        install_mode=install_mode,
-        kometa_root=selection["selected_root"].as_posix() if selection.get("selected_root") else "",
-        kometa_root_display=str(selection["selected_root"]) if selection.get("selected_root") else "",
-        kometa_primary_path=selection["primary_path"].as_posix() if selection.get("primary_path") else "",
-        kometa_primary_path_display=str(selection["primary_path"]) if selection.get("primary_path") else "",
-        kometa_config_dir=selection["config_dir"].as_posix() if selection.get("config_dir") else "",
-        kometa_config_dir_display=str(selection["config_dir"]) if selection.get("config_dir") else "",
-        kometa_log_dir=selection["log_dir"].as_posix() if selection.get("log_dir") else "",
-        kometa_log_dir_display=str(selection["log_dir"]) if selection.get("log_dir") else "",
-        managed_root=managed_root.as_posix(),
-        managed_root_display=str(managed_root),
-        existing_root=existing_root,
-        external_config_root=external_config_root,
-        external_log_root=external_log_root,
-        can_launch=selection.get("can_launch"),
-        can_update=selection.get("can_update"),
-        can_probe_runtime=selection.get("can_probe_runtime"),
-        can_read_logs=selection.get("can_read_logs"),
-    )
-
-
-@app.route("/validate-kometa-root", methods=["POST"])
-def validate_kometa_root():
-    payload = request.get_json(silent=True) or {}
-    logs = []
-
-    def log(msg):
-        print(msg, file=sys.stderr)
-        logs.append(msg)
-
-    target = _resolve_kometa_request_target(payload, logs=logs, require_existing_root=True)
-    if target.get("error"):
-        return jsonify(success=False, error=target["error"], log=logs), 400
-    install_mode = target["install_mode"]
-    p = target["path_obj"]
-    config_dir = target.get("config_dir")
-    log_dir = target.get("log_dir")
-
-    if install_mode == KOMETA_INSTALL_MODE_EXTERNAL:
-        if not p or not p.exists() or not p.is_dir():
-            log("❌ The selected external Kometa config path does not exist in this Quickstart environment.")
-            return jsonify(success=False, error="The selected external Kometa config path does not exist in this Quickstart environment.", log=logs), 400
-        log(f"🔍 Checking external Kometa config path: {p}")
-        if log_dir and Path(log_dir).exists():
-            log(f"📄 External Kometa logs are accessible at: {log_dir}")
-        else:
-            log("ℹ️ External Kometa logs are not currently accessible from this Quickstart environment.")
-        try:
-            sync_result = _sync_generated_yaml_and_assets_to_kometa_config(config_dir, payload.get("config_name", "kometa"), logs=logs)
-        except FileNotFoundError:
-            log("❌ Generated YAML not found.")
-            return jsonify(success=False, error="Generated YAML not found.", log=logs), 500
-        except ValueError as exc:
-            log(f"❌ {exc}")
-            return jsonify(success=False, error=str(exc), log=logs), 400
-        except Exception as exc:
-            log(f"⚠️ Failed to sync config-owned assets referenced in the config: {exc}")
-            return jsonify(success=False, error="Failed to sync generated config to the external Kometa config path.", log=logs), 500
-        log("✅ External Kometa config path is valid and synced.")
-        return (
-            jsonify(
-                success=True,
-                kometa_root="",
-                kometa_root_display="",
-                kometa_config_dir=Path(config_dir).resolve().as_posix() if config_dir else "",
-                kometa_config_dir_display=str(Path(config_dir).resolve()) if config_dir else "",
-                kometa_log_dir=Path(log_dir).resolve().as_posix() if log_dir else "",
-                kometa_log_dir_display=str(Path(log_dir).resolve()) if log_dir else "",
-                venv_python="",
-                venv_python_display="",
-                kometa_version="External / unmanaged",
-                external_mode=True,
-                log=logs,
-                synced_config=str(sync_result.get("destination")),
-            ),
-            200,
-        )
-
-    # Auto-create the Kometa root and config/ if missing
-    if install_mode == KOMETA_INSTALL_MODE_MANAGED and not p.exists():
-        try:
-            p.mkdir(parents=True, exist_ok=True)
-            log(f"📁 Created Kometa root: {p}")
-        except Exception as e:
-            log(f"❌ Failed to create Kometa root: {e}")
-            return jsonify(success=False, error="Failed to create Kometa root.", log=logs), 500
-    elif install_mode == KOMETA_INSTALL_MODE_EXISTING and not p.exists():
-        log("❌ The selected existing Kometa path does not exist in this Quickstart environment.")
-        return jsonify(success=False, error="The selected existing Kometa path does not exist in this Quickstart environment.", log=logs), 400
-    elif install_mode == KOMETA_INSTALL_MODE_EXISTING:
-        missing = _validate_existing_kometa_root(p)
-        if missing:
-            log("❌ The selected existing Kometa path does not look like a Kometa root.")
-            log("ℹ️ Choose the folder that contains kometa.py, requirements.txt, and config/.")
-            return (
-                jsonify(
-                    success=False,
-                    error="Choose the Kometa root folder that contains kometa.py, requirements.txt, and config/.",
-                    log=logs,
-                ),
-                400,
-            )
-
-    try:
-        if install_mode == KOMETA_INSTALL_MODE_MANAGED:
-            (p / "config").mkdir(parents=True, exist_ok=True)
-        elif not (p / "config").exists():
-            log("❌ The selected existing Kometa path is missing its config folder.")
-            return jsonify(success=False, error="The selected existing Kometa path is missing its config folder.", log=logs), 400
-    except Exception as e:
-        log(f"❌ Failed to create config folder: {e}")
-        return jsonify(success=False, error="Failed to create config folder.", log=logs), 500
-
-    # Keep POSIX (internal) and native (display) versions
-    kometa_root_posix = p.as_posix()
-    kometa_root_display = str(p)  # native (Windows => backslashes)
-    session["kometa_root"] = kometa_root_posix  # store normalized internally
-
-    log(f"🔍 Checking path: {kometa_root_display}")
-
-    # --- External tool check (python is required) ---
-    missing_tools = []
-    if shutil.which("python") is None and shutil.which("python3") is None:
-        missing_tools.append("python or python3")
-
-    if missing_tools:
-        for tool in missing_tools:
-            log(f"❌ Required tool not found: {tool}")
-        return jsonify(success=False, error=f"Missing required tools: {', '.join(missing_tools)}", log=logs), 400
-
-    log("✅ All required external tools are available.")
-
-    # Python version (best-effort)
-    try:
-        python_cmd = shutil.which("python") or shutil.which("python3")
-        version_output = subprocess.check_output([python_cmd, "--version"], stderr=subprocess.STDOUT, text=True)
-        log(f"🐍 Detected Python version: {version_output.strip()}")
-    except Exception as e:
-        log(f"⚠️ Failed to detect Python version: {e}")
-
-    # Git version (optional/best-effort)
-    try:
-        git_output = subprocess.check_output(["git", "--version"], stderr=subprocess.STDOUT, text=True)
-        log(f"🔧 Detected Git version: {git_output.strip()}")
-    except Exception as e:
-        log(f"⚠️ Failed to detect Git version: {e}")
-
-    # --- Kometa files check (if you're expecting them to already be present) ---
-    kometa_version = "Unknown"
-    version_path = p / "VERSION"
-    if version_path.exists():
-        try:
-            kometa_version = version_path.read_text(encoding="utf-8").strip()
-            log(f"📦 Kometa version detected: {kometa_version}")
-        except Exception as e:
-            log(f"⚠️ Failed to read VERSION file: {e}")
-
-    required_files = ["kometa.py", "requirements.txt"]
-    for fname in required_files:
-        fpath = p / fname
-        if not fpath.exists():
-            log(f"❌ Required file missing: {fname}")
-            return jsonify(success=False, error=f"{fname} not found.", log=logs), 400
-        log(f"✔️ Found required file: {fname}")
-
-    # --- Virtualenv & deps under <root>/kometa-venv ---
-    is_windows = sys.platform.startswith("win")
-    venv_dir = p / "kometa-venv"
-    bin_dir = venv_dir / ("Scripts" if is_windows else "bin")
-    python_bin = bin_dir / ("python.exe" if is_windows else "python")
-    pip_bin = bin_dir / ("pip.exe" if is_windows else "pip")
-
-    if not venv_dir.exists():
-        log("📦 Creating virtual environment...")
-        try:
-            subprocess.check_call([sys.executable, "-m", "venv", str(venv_dir)])
-            log("✅ Virtual environment created.")
-        except subprocess.CalledProcessError as e:
-            log(f"❌ Failed to create venv: {str(e)}")
-            return jsonify(success=False, error="Failed to create venv.", log=logs), 500
-    else:
-        log("ℹ️ Virtual environment already exists.")
-
-    if not pip_bin.exists():
-        log(f"❌ pip not found in venv at {pip_bin}")
-        return jsonify(success=False, error=f"pip not found in {pip_bin}", log=logs), 500
-
-    log("⬆️ Checking pip version and attempting upgrade...")
-    try:
-        result = subprocess.run([str(python_bin), "-m", "pip", "install", "--upgrade", "pip"], stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True, check=True)
-        output = result.stdout.strip()
-        log("ℹ️ pip is already up to date." if "Requirement already satisfied" in output else "✅ pip upgraded.")
-        for line in output.splitlines():
-            log(f"    {line}")
-    except subprocess.CalledProcessError as e:
-        log(f"❌ pip upgrade failed: {e}")
-        return jsonify(success=False, error="pip upgrade failed.", log=logs), 500
-
-    log("📦 Installing requirements.txt...")
-    try:
-        result = subprocess.run(
-            [str(python_bin), "-m", "pip", "install", "-r", str(p / "requirements.txt")], stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True, check=True
-        )
-        output = result.stdout.strip()
-        log(
-            "ℹ️ All requirements are already satisfied."
-            if "Requirement already satisfied" in output and "Successfully installed" not in output
-            else "✅ requirements.txt installed or updated."
-        )
-        for line in output.splitlines():
-            log(f"    {line}")
-    except subprocess.CalledProcessError as e:
-        log(f"❌ Error installing requirements: {str(e)}")
-        return jsonify(success=False, error="Failed pip install.", log=logs), 500
-
-    try:
-        _sync_generated_yaml_and_assets_to_kometa_config(p / "config", payload.get("config_name", "kometa"), logs=logs)
-    except FileNotFoundError:
-        log("❌ Generated YAML not found.")
-        return jsonify(success=False, error="Generated YAML not found.", log=logs), 500
-    except ValueError as exc:
-        log(f"❌ {exc}")
-        return jsonify(success=False, error=str(exc), log=logs), 400
-    except Exception as e:
-        log(f"⚠️ Failed to sync config-owned assets referenced in the config: {e}")
-
-    log("✅ Kometa root is valid and ready.")
-
-    return (
-        jsonify(
-            success=True,
-            # internal normalized for any future backend use
-            kometa_root=kometa_root_posix,
-            venv_python=python_bin.as_posix(),
-            # native-display for UI/command builder
-            kometa_root_display=kometa_root_display,
-            venv_python_display=str(python_bin),
-            kometa_version=kometa_version,
-            log=logs,
-        ),
-        200,
-    )
-
-
-def _probe_kometa_root_state(path_obj):
-    p = Path(path_obj)
-    kometa_root_posix = p.as_posix()
-    kometa_root_display = str(p)
-    is_windows = sys.platform.startswith("win")
-    venv_dir = p / "kometa-venv"
-    bin_dir = venv_dir / ("Scripts" if is_windows else "bin")
-    python_bin = bin_dir / ("python.exe" if is_windows else "python")
-    version_path = p / "VERSION"
-    kometa_py = p / "kometa.py"
-    requirements = p / "requirements.txt"
-    config_dir = p / "config"
-    version_value = "Unknown"
-    if version_path.exists():
-        try:
-            version_value = version_path.read_text(encoding="utf-8").strip() or "Unknown"
-        except Exception:
-            version_value = "Unknown"
-
-    return {
-        "kometa_root": kometa_root_posix,
-        "kometa_root_display": kometa_root_display,
-        "venv_python": python_bin.as_posix(),
-        "venv_python_display": str(python_bin),
-        "kometa_version": version_value,
-        "root_exists": p.exists(),
-        "config_dir_exists": config_dir.exists(),
-        "kometa_installed": kometa_py.exists() and requirements.exists(),
-        "venv_exists": venv_dir.exists(),
-        "venv_python_exists": python_bin.exists(),
-        "kometa_running": helpers.is_kometa_running(),
-    }
-
-
-@app.route("/probe-kometa-root", methods=["POST"])
-def probe_kometa_root():
-    payload = request.get_json(silent=True) or {}
-    logs = []
-
-    def log(msg):
-        print(msg, file=sys.stderr)
-        logs.append(msg)
-
-    target = _resolve_kometa_request_target(payload, logs=logs, require_existing_root=False)
-    if target.get("error"):
-        return jsonify(success=False, error=target["error"], log=logs), 400
-    if target.get("install_mode") == KOMETA_INSTALL_MODE_EXTERNAL:
-        config_dir = target.get("config_dir")
-        log_dir = target.get("log_dir")
-        state = {
-            "kometa_root": "",
-            "kometa_root_display": "",
-            "kometa_config_dir": Path(config_dir).resolve().as_posix() if config_dir else "",
-            "kometa_config_dir_display": str(Path(config_dir).resolve()) if config_dir else "",
-            "kometa_log_dir": Path(log_dir).resolve().as_posix() if log_dir else "",
-            "kometa_log_dir_display": str(Path(log_dir).resolve()) if log_dir else "",
-            "venv_python": "",
-            "venv_python_display": "",
-            "kometa_version": "External / unmanaged",
-            "root_exists": bool(config_dir and Path(config_dir).exists()),
-            "config_dir_exists": bool(config_dir and Path(config_dir).exists()),
-            "log_dir_exists": bool(log_dir and Path(log_dir).exists()),
-            "kometa_installed": False,
-            "venv_exists": False,
-            "venv_python_exists": False,
-            "kometa_running": False,
-            "external_mode": True,
-        }
-        log(f"🔍 Probing external Kometa config path: {state['kometa_config_dir_display']}")
-        if state["config_dir_exists"]:
-            log("✅ External Kometa config path is accessible.")
-        else:
-            log("❌ External Kometa config path is not currently accessible.")
-        if state["log_dir_exists"]:
-            log(f"📄 External Kometa logs are accessible at: {state['kometa_log_dir_display']}")
-        else:
-            log("ℹ️ External Kometa log path is not currently accessible.")
-        log("ℹ️ Quickstart cannot probe or launch the external Kometa runtime directly in this mode.")
-        return jsonify(success=True, log=logs, **state), 200
-    if target.get("install_mode") == KOMETA_INSTALL_MODE_EXISTING:
-        missing = _validate_existing_kometa_root(target["path_obj"])
-        if missing:
-            log("❌ The selected existing Kometa path does not look like a Kometa root.")
-            log("ℹ️ Choose the folder that contains kometa.py, requirements.txt, and config/.")
-            return (
-                jsonify(
-                    success=False,
-                    error="Choose the Kometa root folder that contains kometa.py, requirements.txt, and config/.",
-                    log=logs,
-                ),
-                400,
-            )
-    p = target["path_obj"]
-
-    state = _probe_kometa_root_state(p)
-    log(f"🔍 Probing Kometa path: {state['kometa_root_display']}")
-    if not state["root_exists"]:
-        log("ℹ️ Kometa root does not exist yet. Install required.")
-    elif not state["kometa_installed"]:
-        log("ℹ️ Kometa files not found yet. Install required.")
-    else:
-        log("✅ Kometa files detected locally.")
-        if state["kometa_version"]:
-            log(f"📦 Local Kometa version: {state['kometa_version']}")
-        if state["venv_python_exists"]:
-            log(f"🐍 Kometa venv python detected at: {state['venv_python_display']}")
-        else:
-            log("ℹ️ Kometa venv python not present yet. Prepare step still needed.")
-    if state["kometa_running"]:
-        log("ℹ️ Kometa is currently running.")
-
-    return jsonify(success=True, log=logs, **state), 200
-
-
-@app.route("/check-kometa-update", methods=["POST"])
-def check_kometa_update():
-    payload = request.get_json(silent=True) or {}
-    branch_override_raw = payload.get("branch_override")
-    branch_override = helpers.normalize_kometa_branch_override(branch_override_raw)
-    logs = []
-
-    def log(msg):
-        print(msg, file=sys.stderr)
-        logs.append(msg)
-
-    if branch_override_raw and not branch_override:
-        log(f"❌ Invalid Kometa branch override: {branch_override_raw}")
-        return jsonify(success=False, error="Invalid Kometa branch override.", log=logs), 400
-
-    target = _resolve_kometa_request_target(payload, logs=logs, require_existing_root=False)
-    if target.get("error"):
-        return jsonify(success=False, error=target["error"], log=logs), 400
-    if target.get("install_mode") == KOMETA_INSTALL_MODE_EXTERNAL:
-        config_dir = target.get("config_dir")
-        log_dir = target.get("log_dir")
-        log("ℹ️ Quickstart cannot check or update an external Kometa runtime in config/log-only mode.")
-        return (
-            jsonify(
-                success=True,
-                log=logs,
-                update_check_completed=False,
-                kometa_update_check_skipped=True,
-                local_version="External / unmanaged",
-                remote_version="",
-                kometa_update_available=False,
-                cached=False,
-                kometa_installed=False,
-                kometa_running=False,
-                external_mode=True,
-                kometa_root="",
-                kometa_root_display="",
-                kometa_config_dir=Path(config_dir).resolve().as_posix() if config_dir else "",
-                kometa_config_dir_display=str(Path(config_dir).resolve()) if config_dir else "",
-                kometa_log_dir=Path(log_dir).resolve().as_posix() if log_dir else "",
-                kometa_log_dir_display=str(Path(log_dir).resolve()) if log_dir else "",
-            ),
-            200,
-        )
-    if target.get("install_mode") == KOMETA_INSTALL_MODE_EXISTING:
-        missing = _validate_existing_kometa_root(target["path_obj"])
-        if missing:
-            log("❌ The selected existing Kometa path does not look like a Kometa root.")
-            log("ℹ️ Choose the folder that contains kometa.py, requirements.txt, and config/.")
-            return (
-                jsonify(
-                    success=False,
-                    error="Choose the Kometa root folder that contains kometa.py, requirements.txt, and config/.",
-                    log=logs,
-                ),
-                400,
-            )
-    p = target["path_obj"]
-
-    state = _probe_kometa_root_state(p)
-    if not state["kometa_installed"]:
-        log("ℹ️ Kometa is not installed yet; update check skipped.")
-        return (
-            jsonify(
-                success=True,
-                log=logs,
-                update_check_completed=False,
-                kometa_update_check_skipped=False,
-                local_version=state["kometa_version"],
-                remote_version="",
-                kometa_update_available=False,
-                cached=False,
-                **state,
-            ),
-            200,
-        )
-
-    if state["kometa_running"]:
-        log("ℹ️ Kometa is currently running; update check skipped.")
-        return (
-            jsonify(
-                success=True,
-                log=logs,
-                update_check_completed=True,
-                kometa_update_check_skipped=True,
-                local_version=state["kometa_version"],
-                remote_version="",
-                kometa_update_available=False,
-                cached=False,
-                **state,
-            ),
-            200,
-        )
-
-    if branch_override:
-        log(f"⚠️ Kometa branch override selected: {branch_override}")
-    else:
-        log("ℹ️ Kometa branch selection: auto")
-
-    update_info = helpers.get_cached_kometa_update(
-        p,
-        force_refresh=helpers.booler(payload.get("force", False)),
-        branch_override=branch_override,
-    )
-    local_version = update_info.get("local_version") or state["kometa_version"]
-    remote_version = update_info.get("remote_version") or ""
-    remote_branch = update_info.get("branch") or "nightly"
-    local_branch = update_info.get("local_branch") or "unknown"
-    local_sha = update_info.get("local_sha") or ""
-    remote_sha = update_info.get("remote_sha") or ""
-    comparison_basis = update_info.get("comparison_basis") or "version"
-    remote_version_url = helpers.GITHUB_BASE_URL + f"/{remote_branch}/VERSION"
-    log(f"🌐 Remote VERSION source: {remote_version_url}")
-    log(f"ℹ️ Installed Kometa branch metadata: {local_branch}")
-    if local_sha:
-        log(f"🔎 Local Kometa SHA: {local_sha[:12]}")
-    if remote_sha:
-        log(f"🔎 Remote Kometa SHA: {remote_sha[:12]}")
-    log(f"ℹ️ Update comparison basis: {comparison_basis}")
-    if update_info.get("cached"):
-        log("ℹ️ Using cached Kometa update lookup.")
-    if update_info.get("update_available"):
-        if update_info.get("branch_mismatch"):
-            log(f"⚠️ Installed branch '{local_branch}' differs from selected branch '{remote_branch}'.")
-        log(f"⬆️ Update available: {local_version} → {remote_version}")
-    else:
-        log(f"✅ Kometa is up to date: {local_version}")
-
-    return (
-        jsonify(
-            success=True,
-            log=logs,
-            update_check_completed=True,
-            kometa_update_check_skipped=False,
-            local_version=local_version,
-            remote_version=remote_version,
-            kometa_update_available=bool(update_info.get("update_available")),
-            cached=bool(update_info.get("cached")),
-            **state,
-        ),
-        200,
-    )
-
-
-@app.route("/update-kometa", methods=["POST"])
-def update_kometa():
-    blocker = _get_active_work_blocker("kometa_update")
-    if blocker:
-        pid = blocker.get("pid")
-        target_page = blocker.get("target_page")
-        return (
-            jsonify(
-                {
-                    "success": False,
-                    "error": f"Kometa is currently running (PID {pid}). Stop it before updating.",
-                    "blocked_by": blocker.get("blocked_by"),
-                    "pid": pid,
-                    "target_page": target_page,
-                }
-            ),
-            409,
-        )
-    try:
-        data = request.get_json(silent=True) or {}
-        target = _resolve_kometa_request_target(data, require_existing_root=True)
-        if target.get("error"):
-            return jsonify({"success": False, "error": target["error"], "log": [f"❌ {target['error']}"]}), 400
-        if target.get("install_mode") == KOMETA_INSTALL_MODE_EXTERNAL:
-            return (
-                jsonify(
-                    {
-                        "success": False,
-                        "error": "External Kometa mode does not support runtime updates. Change to a managed or existing direct install on the Start page to update from Quickstart.",
-                        "log": [
-                            "❌ External Kometa mode does not support runtime updates.",
-                            "ℹ️ Change to a managed or existing direct install on the Start page if Quickstart should update Kometa itself.",
-                        ],
-                    }
-                ),
-                400,
-            )
-        if target.get("install_mode") == KOMETA_INSTALL_MODE_EXISTING:
-            return (
-                jsonify(
-                    {
-                        "success": False,
-                        "error": "Existing direct Kometa installs must be updated manually outside Quickstart. Quickstart can validate and check version status, but it will not modify that install.",
-                        "log": [
-                            "❌ Existing direct Kometa installs are manual-update only in Quickstart.",
-                            "ℹ️ Quickstart can validate and check version status for this install, but update it manually outside Quickstart to reduce the risk of runtime failures.",
-                        ],
-                    }
-                ),
-                400,
-            )
-        kometa_root = target["path_obj"]
-        install_mode = target["install_mode"]
-        branch_override_raw = data.get("branch_override")
-        branch_override = helpers.normalize_kometa_branch_override(branch_override_raw)
-        if branch_override_raw and not branch_override:
-            return jsonify({"success": False, "error": "Invalid Kometa branch override.", "log": ["❌ Invalid Kometa branch override."]}), 400
-        qs_branch = data.get("branch") or helpers.detect_git_branch(app.root_path)
-        kometa_branch = branch_override or ("master" if qs_branch == "master" else "nightly")
-        force_update = helpers.booler(data.get("force", False))
-        background = data.get("background") is True
-
-        if background:
-            active = _get_active_background_job("kometa_update")
-            if active:
-                phase = active.get("phase")
-                if phase and phase not in ["done", "error"]:
-                    return (
-                        jsonify(success=True, active=True, existing_job=True, job_id=active.get("job_id"), phase=phase),
-                        200,
-                    )
-                _clear_active_background_job("kometa_update", job_id=active.get("job_id"))
-
-            job = _create_background_job(
-                "kometa_update",
-                trigger="manual",
-                phase="queued",
-                status="running",
-                target_page=JOB_TARGET_PAGES.get("kometa_update"),
-                logs=[],
-                done=False,
-                success=False,
-                up_to_date=False,
-                skipped=False,
-                force=force_update,
-                qs_branch=qs_branch,
-                kometa_branch=kometa_branch,
-                started_epoch=time.time(),
-            )
-            job_id = job["job_id"]
-
-            def worker():
-                class _ProgressLog(list):
-                    def append(self_inner, item):
-                        super().append(item)
-                        current = _get_background_job(job_id) or {}
-                        lines = list(current.get("logs") or [])
-                        lines.append(item)
-                        _update_background_job(job_id, logs=lines)
-
-                logs = _ProgressLog()
-                _update_background_job(job_id, phase="running", status="running")
-                logs.append(f"🔎 Quickstart branch: {qs_branch}")
-                logs.append(f"📍 Kometa install mode: {install_mode}")
-                logs.append(f"📁 Target Kometa root: {kometa_root}")
-                if branch_override:
-                    logs.append(f"⚠️ Kometa branch override selected: {branch_override}")
-                else:
-                    logs.append("ℹ️ Kometa branch selection: auto")
-                logs.append(f"⚙️ Kometa branch selected: {kometa_branch} (ZIP mode)")
-                if force_update:
-                    logs.append("Force update enabled.")
-
-                try:
-                    if install_mode == KOMETA_INSTALL_MODE_EXISTING:
-                        result = helpers.perform_kometa_update_zip_only_at_root(kometa_root, branch=kometa_branch, force=force_update, logs=logs)
-                    else:
-                        result = helpers.perform_kometa_update_zip_only(helpers.CONFIG_DIR, branch=kometa_branch, force=force_update, logs=logs)
-                    try:
-                        invalidate_target = kometa_root if install_mode == KOMETA_INSTALL_MODE_EXISTING else helpers.CONFIG_DIR
-                        helpers.invalidate_cached_kometa_update(invalidate_target)
-                    except Exception:
-                        pass
-                    if result.get("success", False):
-                        _complete_background_job(
-                            job_id,
-                            phase="done",
-                            success=True,
-                            done=True,
-                            up_to_date=bool(result.get("up_to_date", False)),
-                            skipped=bool(result.get("skipped", False)),
-                        )
-                    else:
-                        _fail_background_job(
-                            job_id,
-                            "Kometa update failed.",
-                            done=True,
-                            success=False,
-                            up_to_date=bool(result.get("up_to_date", False)),
-                            skipped=bool(result.get("skipped", False)),
-                        )
-                except Exception as e:
-                    logs.append("Exception during Kometa update.")
-                    helpers.ts_log(f"Kometa update failed: {e}", level="ERROR")
-                    _fail_background_job(job_id, e, done=True, success=False)
-                finally:
-                    _clear_active_background_job("kometa_update", job_id=job_id)
-
-            threading.Thread(target=worker, daemon=True).start()
-            return jsonify(success=True, active=True, job_id=job_id, phase="queued"), 200
-
-        logs = []
-        logs.append(f"🔎 Quickstart branch: {qs_branch}")
-        logs.append(f"📍 Kometa install mode: {install_mode}")
-        logs.append(f"📁 Target Kometa root: {kometa_root}")
-        if branch_override:
-            logs.append(f"⚠️ Kometa branch override selected: {branch_override}")
-        else:
-            logs.append("ℹ️ Kometa branch selection: auto")
-        logs.append(f"⚙️ Kometa branch selected: {kometa_branch} (ZIP mode)")
-        if force_update:
-            logs.append("Force update enabled.")
-
-        if install_mode == KOMETA_INSTALL_MODE_EXISTING:
-            result = helpers.perform_kometa_update_zip_only_at_root(kometa_root, branch=kometa_branch, force=force_update, logs=logs)
-        else:
-            result = helpers.perform_kometa_update_zip_only(helpers.CONFIG_DIR, branch=kometa_branch, force=force_update, logs=logs)
-        try:
-            invalidate_target = kometa_root if install_mode == KOMETA_INSTALL_MODE_EXISTING else helpers.CONFIG_DIR
-            helpers.invalidate_cached_kometa_update(invalidate_target)
-        except Exception:
-            pass
-        status = 200 if result.get("success") else 500
-
-        return (
-            jsonify(
-                {
-                    "success": result.get("success", False),
-                    "log": list(logs),
-                    "qs_branch": qs_branch,
-                    "kometa_branch": kometa_branch,
-                    "up_to_date": result.get("up_to_date", False),
-                    "skipped": result.get("skipped", False),
-                    "force": force_update,
-                }
-            ),
-            status,
-        )
-
-    except Exception as e:
-        helpers.ts_log(f"Kometa update failed: {e}", level="ERROR")
-        return jsonify({"success": False, "log": ["Exception during Kometa update."]}), 500
-
-
-@app.route("/update-kometa-progress", methods=["GET"])
-def update_kometa_progress():
-    job_id = request.args.get("job_id", "").strip()
-    since = request.args.get("since", "0").strip()
-    if not job_id:
-        return jsonify(success=False, error="Missing job_id."), 400
-    info = _get_background_job(job_id)
-    if not info:
-        return jsonify(success=False, error="Unknown job_id."), 404
-    try:
-        start_idx = max(int(since or "0"), 0)
-    except ValueError:
-        start_idx = 0
-    lines = list(info.get("logs") or [])
-    return jsonify(
-        success=True,
-        job_id=job_id,
-        phase=info.get("phase"),
-        done=bool(info.get("done")) or info.get("status") in {"complete", "error"},
-        update_success=bool(info.get("success")),
-        up_to_date=bool(info.get("up_to_date")),
-        skipped=bool(info.get("skipped")),
-        force=bool(info.get("force")),
-        qs_branch=info.get("qs_branch"),
-        kometa_branch=info.get("kometa_branch"),
-        lines=lines[start_idx:],
-        next_index=len(lines),
-    )
-
-
 def _probe_imagemaid_root_state(path_obj):
     p = Path(path_obj)
     imagemaid_root_posix = p.as_posix()
@@ -15783,21 +14452,6 @@ def _imagemaid_settings_to_form_payload(payload):
         if key in payload:
             form_payload[f"imagemaid_{key}"] = payload.get(key)
     return form_payload
-
-
-def _retrieve_settings_for_config(config_name, target):
-    source, source_name = persistence.extract_names(target)
-    stored_validated, stored_user_entered, stored_payload = database.retrieve_section_data(config_name, source_name)
-    payload = stored_payload if isinstance(stored_payload, dict) else {}
-    section = payload.get(source_name, {}) if isinstance(payload.get(source_name), dict) else {}
-    if not section:
-        section = persistence.get_dummy_data(source_name)
-    return {
-        "validated": helpers.booler(stored_validated),
-        "user_entered": helpers.booler(stored_user_entered),
-        "validated_at": payload.get("validated_at") if isinstance(payload, dict) else None,
-        source_name: section,
-    }
 
 
 def _get_stored_plex_credentials_for_config(config_name):
